Use paste.deploy to parse alembic config files
Alembic uses configparser to parse its config files, which is
compatible with the paste.deploy ini config files in terms of
encoding, but paste.deploy has additional features such as
`user = config:` lines. This reduces the amount of places that
alembic config files are generated throughout the codebase and
uses paste.deploy to extract the config values for the `app:conduit`
section before directly inserting them into an alembic config.

This allows users to omit `script_location` from test.ini, and moreover
to omit any parameters required by alembic that are set in a parent
of the relevant app:conduit block.