Split alembic configuration into alembic.ini
This change decouples the pyramid and alembic config ini files,
as it's normal for the alembic config to be common across all
deployment types, but the alembic config parser does not support
ini file inheritence.

This uses the alembic -x mechanism to pass the pyramid config files
into the alembic script when it's being run directly, i.e.
alembic -x ini=etc/development.ini ... and hard-codes `alembic.ini`
as the location of the alembic config when Pyramid invokes alembic.
This is the default location for alembic files, and removes the
requirement to add a `-c` parameter to alembic invocations.

The alembic env file which is loaded by the alembic script will check
to see if a Base.metadata is already configured with a bind, and if not
the sqlalchemy.url and sqlalchemy.echo parameters are used directly
from the registry after calling boostrap() on the ini file from the x
parameter.

If Base.metadata is set, that means the env is being invoked from within
a Pyramid context, and the registry is introspected without first bootstrapping.