Updare circleci to try both development and production ini files.
Previously, only pshell development.ini was tested, not the production
ini file. In addition, as the deferred sqlalchemy library is being used
this would not actually check that the database was connectable. The
on-the-fly exit.sh script has been replaced with a static one that
attempts to connect to the database and returns an appropriate
exit code.